refactor ModelPart to three clases
`Feedable` has feed dict method, batch size and train mode placeholders.
perhaps will contain an easy method of adding placeholders to graph
hopefully, will be useful when refactoring get_dependencies

`Parameterized` manages the model part functionality around variables:
scopes, initializers, and saving/loading

`ModelPart` to be further refactored, only functionality that remains
here is the get_dependencies function.
